requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
open_clip_torch>=2.20.0
ftfy>=6.1.1
regex>=2023.0.0

# Backend
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
httpx>=0.24.0

# Data processing
scikit-learn>=1.2.0
matplotlib>=3.7.0
tqdm>=4.65.0

# Development
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
